+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added
 
+- Issue #54: Modify edge with mouse
 - Issue #156: Curve modules
 
 ## [4.3.3] - 2021-11-25